+++ XRViewer/AppStateController/AppState.h
@@ -79,7 +79,6 @@ typedef NS_OPTIONS(NSUInteger, ShowOptions)
 @property(nonatomic) ShowMode showMode;
 @property(nonatomic) ShowOptions showOptions;
 @property(nonatomic) BOOL webXR;
-@property(nonatomic) BOOL interruption;
 @property(nonatomic) BOOL computerVisionFrameRequested;
 @property(nonatomic) BOOL shouldRemoveAnchorsOnNextARSession;
 @property(nonatomic) BOOL sendComputerVisionData;
@@ -96,6 +95,5 @@ typedef NS_OPTIONS(NSUInteger, ShowOptions)
 - (instancetype)updatedShowOptions:(ShowOptions)showOptions;
 - (instancetype)updatedWebXR:(BOOL)webXR;
 - (instancetype)updatedWithARRequest:(NSDictionary *)dict;
-- (instancetype)updatedWithInterruption:(BOOL)interruption;
 
 @end